fairsim-app/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
shap==0.45.1
imbalanced-learn==0.12.3
google-generativeai==0.7.2
python-dotenv==1.0.1
joblib==1.4.2
pydantic==2.7.1
chardet==5.2.0